+++ xenon-common/src/test/java/com/vmware/xenon/common/opentracing/TracerFactoryTest.java
@@ -14,12 +14,15 @@
 package com.vmware.xenon.common.opentracing;
 
 import static junit.framework.TestCase.assertEquals;
+import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
 
 import brave.opentracing.BraveTracer;
 import io.opentracing.NoopTracer;
 import io.opentracing.Tracer;
+import org.junit.Rule;
 import org.junit.Test;
+import org.junit.rules.ExternalResource;
 
 import com.vmware.xenon.common.ServiceHost;
 import com.vmware.xenon.common.config.TestXenonConfiguration;
@@ -27,12 +30,16 @@ import com.vmware.xenon.common.test.VerificationHost;
 
 public class TracerFactoryTest {
 
+    @Rule
+    public InjectCleanFactory injectCleanFactory = new InjectCleanFactory();
+
     @Test
     public void defaultIsNoop() throws Exception {
         TestXenonConfiguration.override(TracerFactory.class, "provider", null);
         ServiceHost h = VerificationHost.create(0);
         Tracer tracer = TracerFactory.factory.create(h);
         assertTrue(tracer instanceof NoopTracer);
+        assertFalse(TracerFactory.factory.enabled());
     }
 
     @Test
@@ -42,6 +49,7 @@ public class TracerFactoryTest {
         ServiceHost h = VerificationHost.create(0);
         Tracer tracer = TracerFactory.factory.create(h);
         assertTrue(tracer instanceof com.uber.jaeger.Tracer);
+        assertTrue(TracerFactory.factory.enabled());
     }
 
     @Test
@@ -63,6 +71,7 @@ public class TracerFactoryTest {
         ServiceHost h = VerificationHost.create(0);
         Tracer tracer = TracerFactory.factory.create(h);
         assertTrue(tracer instanceof BraveTracer);
+        assertTrue(TracerFactory.factory.enabled());
     }
 
     @Test(expected = RuntimeException.class)
@@ -71,4 +80,22 @@ public class TracerFactoryTest {
         ServiceHost h = VerificationHost.create(0);
         TracerFactory.factory.create(h);
     }
+
+    public class InjectCleanFactory extends ExternalResource {
+        TracerFactory factory;
+
+        @Override
+        protected void after() {
+            TracerFactory.factory = this.factory;
+            super.after();
+        }
+
+        @Override
+        protected void before() throws Throwable {
+            super.before();
+            this.factory = TracerFactory.factory;
+            TracerFactory.factory = new TracerFactory();
+        }
+    }
+
 }